backend/requirements.txt
# ===========================================
# MOA Backend Dependencies
# ===========================================

# FastAPI & Server
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9

# Database
sqlalchemy[asyncio]==2.0.25
asyncpg==0.29.0
alembic==1.13.1

# Validation & Serialization
pydantic==2.6.1
pydantic-settings==2.1.0
email-validator==2.1.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Redis & Task Queue
redis==5.0.1
celery==5.3.6

# Storage (S3/MinIO)
boto3==1.34.34

# HTTP Client
httpx==0.26.0

# AI/LLM
anthropic==0.18.1

# Utilities
python-dotenv==1.0.1
orjson==3.9.13

# Testing
pytest==8.0.0
pytest-asyncio==0.23.4

# Development
ruff==0.2.1
